feat: Company Registry and Saudi Exchange sync
This module performs periodic check with Saudi Exchange,
to to stay in sync with the latest symbols. As this database, might
contains delisted companies, this module is responsible for this task.

Of course not "deleting" the delisted company from the database,
just make it "inaccessible" to the user.

What happens in the cloud, stays in the cloud.